requirements.txt
sphinx
sphinx-sitemap
sphinx-copybutton
sphinx-design
sphinx_sqlalchemy
ipykernel
jupyter-sphinx
sphinx-togglebutton
sphinxcontrib-mermaid
sphinx-docsearch
nbsphinx
myst-parser
numpy
numpydoc
matplotlib
pandas
Babel
